+++ test/cheerio_crawler.js
@@ -7,7 +7,6 @@ import rqst from 'request';
 import chai, { expect } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import log from 'apify-shared/log';
-import _ from 'underscore';
 import { delayPromise } from 'apify-shared/utilities';
 import { ENV_VARS } from 'apify-shared/consts';
 import Apify from '../build/index';
@@ -118,62 +117,6 @@ describe('CheerioCrawler', () => {
             expect(request.userData.title).to.be.eql(String(id));
             expect(request.userData.html).to.be.a('string');
             expect(request.userData.html.length).not.to.be.eql(0);
-        });
-    });
-
-    it('should abort and resume', async () => {
-        const sources = _.range(100).map(index => ({ url: `https://example.com/?q=${index + 1}` }));
-        let cheerioCrawler;
-        let isStopped = false;
-        const processed = [];
-        const failed = [];
-        const requestList = new Apify.RequestList({ sources });
-        const requestFunction = async ({ request }) => {
-            await delayPromise(2);
-            return `<html><head><title>${request.url[request.url.length - 1]}</title></head><body>Body</body></html>`;
-        };
-        const handlePageFunction = async ({ $, html, request }) => {
-            if (request.url.endsWith('45') && !isStopped) {
-                await cheerioCrawler.abort();
-                isStopped = true;
-            } else {
-                request.userData.title = $('title').text();
-                request.userData.html = html;
-                processed.push(request);
-            }
-        };
-
-        cheerioCrawler = new Apify.CheerioCrawler({
-            requestList,
-            minConcurrency: 5,
-            maxConcurrency: 5,
-            requestFunction,
-            handlePageFunction,
-            handleFailedRequestFunction: ({ request }) => failed.push(request),
-        });
-
-        await requestList.initialize();
-        await cheerioCrawler.run();
-
-        expect(processed.length).to.be.within(40, 50);
-        expect(failed).to.have.lengthOf(0);
-
-        processed.sort(comparator);
-
-        for (let i = 0; i < 40; i++) {
-            const request = processed[i];
-            expect(request.url).to.be.eql(sources[i].url);
-            expect(request.userData.title).to.be.eql(request.url[request.url.length - 1]);
-        }
-
-        await Apify.utils.sleep(10); // Wait for event loop to unwind.
-        await cheerioCrawler.run();
-
-        expect(processed.length).to.be.within(100, 110);
-        expect(failed).to.have.lengthOf(0);
-        expect(new Set(processed.map(p => p.url))).to.be.eql(new Set(sources.map(s => s.url)));
-        processed.forEach((request) => {
-            expect(request.userData.title).to.be.eql(request.url[request.url.length - 1]);
         });
     });