+++ test/request_queue.js
@@ -1,7 +1,6 @@
 import chai, { assert, expect } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import _ from 'underscore';
-import 'babel-polyfill';
 import sinon from 'sinon';
 import path from 'path';
 import { ENV_VARS } from 'apify-shared/consts';